[spaces.dg.femlocalfunctions] moved DiscontinuousLangrange -> DG, refs #14
